feat(onchain_endpoint): Removed onchain oracle calls
