+++ pragma-node/src/infra/mod.rs
@@ -1,3 +1,2 @@
 pub mod kafka;
-pub mod onchain;
 pub mod repositories;
