--- pragma-node/src/infra/onchain/mod.rs
@@ -1,1 +1,0 @@
-pub mod oracle;
